Conclusão da aula 2, refatoração do contexto, modelagem do hook de effect, criação do hook useAuth
