+++ public/index.html
@@ -6,7 +6,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&family=Roboto:wght@400;700&display=swap" rel="stylesheet">
-    <title>Fraud test APP</title>
+    <title>Let me Ask</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
