+++ src/App.tsx
@@ -1,50 +1,17 @@
-import { createContext, useState } from 'react';
 import {BrowserRouter, Route} from 'react-router-dom'
+import { AuthContextProvider } from './context/AuthContext';
 
 import { Home } from "./pages/Home";
 import { NewRoom } from './pages/NewRoom';
-import { auth, firebase } from './services/firebase';
-
-type User = {
-  id: string;
-  name: string;
-  avatar: string | null;
-}
-
-type AuthContextType = {
-  user: User | undefined;
-  signInWithGoogle: () => Promise<void>;
-}
-
-export const AuthContext = createContext({} as AuthContextType);
 
 function App() {
-  const [user, setUser] = useState<User>();
-
-  async function signInWithGoogle() {
-    const provider = new firebase.auth.GoogleAuthProvider();
-
-    const result = await auth.signInWithPopup(provider);
-
-      if (result.user) {
-        const {displayName, photoURL, uid } = result.user
-        if(!displayName) {
-          throw new Error('Missing information from google Account.');
-        }
-        setUser({
-          id: uid,
-          name: displayName,
-          avatar: photoURL
-        })
-
-      }
-  }
+  
   return (
     <BrowserRouter>
-      <AuthContext.Provider value={{user, signInWithGoogle}}>
+      <AuthContextProvider>
         <Route path="/" exact component={Home} />
         <Route path="/room/new" component={NewRoom} />
-      </AuthContext.Provider>
+      </AuthContextProvider>
     </BrowserRouter>
     
   );
