+++ src/context/AuthContext.tsx
@@ -1,0 +1,68 @@
+import {createContext, ReactNode, useEffect, useState} from "react"
+import {auth, firebase} from "../services/firebase"
+
+type User = {
+        id: string;
+        name: string;
+        avatar: string | null;
+    }
+  
+type AuthContextType = {
+        user: User | undefined;
+        signInWithGoogle: () => Promise<void>;
+    }
+
+type AuthContextProviderProps = {
+        children: ReactNode;
+    }
+
+export const AuthContext = createContext({} as AuthContextType);
+
+export function AuthContextProvider(props:AuthContextProviderProps ){
+
+    const [user, setUser] = useState<User>();
+
+    useEffect(() => {
+        const unsubscribe =  auth.onAuthStateChanged(user => {
+        if(user){
+            const {displayName, photoURL, uid } = user
+            if(!displayName) {
+            throw new Error('Missing information from google Account.');
+            }
+            setUser({
+            id: uid,
+            name: displayName,
+            avatar: photoURL
+            })
+        }
+        })
+        return () => {
+        unsubscribe();
+        }
+    }, [])
+
+    async function signInWithGoogle() {
+        const provider = new firebase.auth.GoogleAuthProvider();
+
+        const result = await auth.signInWithPopup(provider);
+
+        if (result.user) {
+            const {displayName, photoURL, uid } = result.user
+            if(!displayName) {
+            throw new Error('Missing information from google Account.');
+            }
+            setUser({
+            id: uid,
+            name: displayName,
+            avatar: photoURL
+            })
+
+        }
+    }
+    return (
+        <AuthContext.Provider value={{user, signInWithGoogle}}>
+            {props.children}
+        </AuthContext.Provider>
+    )
+
+}
