+++ src/pages/Home.tsx
@@ -4,14 +4,12 @@ import logoImg from '../assets/images/logo.svg'
 import googleIconImg from '../assets/images/google-icon.svg'
 import { Button } from '../components/Button'
 import {auth, firebase} from '../services/firebase'
-
 import '../styles/auth.scss'
-import { useContext } from 'react'
-import { AuthContext } from '../App'
+import { useAuth } from '../hooks/useAuth'
 
 export function Home() {
     const history = useHistory();
-    const {user, signInWithGoogle} = useContext(AuthContext)
+    const {user, signInWithGoogle} = useAuth()
 
     async function handleCreateRoom() {
         if (!user) {
