+++ src/pages/NewRoom.tsx
@@ -1,14 +1,13 @@
-import { useContext } from 'react'
 import { Link } from 'react-router-dom'
-import { AuthContext } from '../App'
 import illustrationImg from '../assets/images/illustration.svg'
 import logoImg from '../assets/images/logo.svg'
 import { Button } from '../components/Button'
 
 import '../styles/auth.scss'
+import { useAuth } from '../hooks/useAuth'
 
 export function NewRoom() {
-    const {user} = useContext(AuthContext)
+    const {user} = useAuth()
     return(
         <div id="page-auth">
             <aside>
